virtio: add interface to enable interrupts
This just adds the plumbing in the virito library.  The next patch will
add support in virito_pci.

The interface allows allocating a single interrupt vector for all
virtqueues.  Normally this wouldn't be very useful, but there are some
devices that won't initialize unless MSI-X is configured, so this will
allow us to use them.

Change-Id: I90602f79a340b8de74f484f843ea2beb8b5d977b
Signed-off-by: Konrad Sztyber <ksztyber@nvidia.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/28397
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Reviewed-by: Ben Walker <ben@nvidia.com>
Community-CI: Mellanox Build Bot